backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
pymupdf
python-docx
python-dotenv
langchain
langchain-ollama
langchain-huggingface
sentence-transformers
langchain-chroma
hwp2yaml
rank-bm25
reportlab
